+++ camera_ops/__init__.py
@@ -2,6 +2,7 @@ from . import (cam_data,
                cam_toggle_parent,
                cam_quick_track_to,
                cam_create,
+               cam_exclude_filter,
                )
 
 modules = (
@@ -9,6 +10,7 @@ modules = (
     cam_toggle_parent,
     cam_quick_track_to,
     cam_create,
+    cam_exclude_filter,
 )
 
 def register():